congruence: make Congruence a class template
Compiles but doesn't link
