+++ tests/cong.test.cc
@@ -18,6 +18,7 @@
 
 #include <utility>
 
+#include "../src/bmat.h"
 #include "../src/cong.h"
 #include "catch.hpp"
 
@@ -31,7 +32,7 @@ TEST_CASE("Congruence 00: 5-parameter constructor",
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong("left", 2, rels, extra);
+  Congruence<> cong("left", 2, rels, extra);
 }
 
 TEST_CASE("Congruence 01: Small fp semigroup",
@@ -41,7 +42,7 @@ TEST_CASE("Congruence 01: Small fp semigroup",
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong("twosided", 2, rels, extra);
+  Congruence<> cong("twosided", 2, rels, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(!cong.is_done());
@@ -69,7 +70,7 @@ TEST_CASE("Congruence 02: Small left congruence on free semigroup",
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong("left", 2, rels, extra);
+  Congruence<> cong("left", 2, rels, extra);
   cong.set_report(CONG_REPORT);
   REQUIRE(cong.nr_classes() == 5);
 }
@@ -81,7 +82,7 @@ TEST_CASE("Congruence 03: Small right congruence on free semigroup",
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong("right", 2, rels, extra);
+  Congruence<> cong("right", 2, rels, extra);
   cong.set_report(CONG_REPORT);
   REQUIRE(cong.nr_classes() == 5);
   REQUIRE(cong.is_done());
@@ -95,7 +96,7 @@ TEST_CASE(
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong("left", 2, rels, extra);
+  Congruence<> cong("left", 2, rels, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.word_to_class_index({0, 0, 1})
@@ -115,7 +116,7 @@ TEST_CASE("Congruence 05: word_to_class_index for small fp semigroup",
   rels.push_back(relation_t({0}, {1, 1}));     // (a, b^2)
   std::vector<relation_t> extra;
 
-  Congruence cong1("twosided", 2, rels, extra);
+  Congruence<> cong1("twosided", 2, rels, extra);
   cong1.set_report(CONG_REPORT);
 
   REQUIRE(cong1.word_to_class_index({0, 0, 1})
@@ -126,7 +127,7 @@ TEST_CASE("Congruence 05: word_to_class_index for small fp semigroup",
           == cong1.word_to_class_index({1, 1}));
   REQUIRE(cong1.word_to_class_index({1}) != cong1.word_to_class_index({0}));
 
-  Congruence cong2("twosided", 2, rels, extra);
+  Congruence<> cong2("twosided", 2, rels, extra);
   cong2.set_report(CONG_REPORT);
 
   REQUIRE(cong2.word_to_class_index({0, 0, 0, 0})
@@ -146,7 +147,7 @@ TEST_CASE("Congruence 06: 6-argument constructor (trivial cong)",
   REQUIRE(S.degree() == 5);
 
   std::vector<relation_t> extra;
-  Congruence              cong("twosided", &S, extra);
+  Congruence<>            cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
   REQUIRE(!cong.is_done());
 
@@ -171,7 +172,7 @@ TEST_CASE("Congruence 07: 6-argument constructor (nontrivial cong)",
   S.factorisation(w1, S.position(t1));
   S.factorisation(w2, S.position(t2));
   std::vector<relation_t> extra({relation_t(w1, w2)});
-  Congruence              cong("twosided", &S, extra);
+  Congruence<>            cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
 
   t1->really_delete();
@@ -196,7 +197,7 @@ TEST_CASE("Congruence 8T: transformation semigroup size 88",
   S.factorisation(w1, S.position(t1));
   S.factorisation(w2, S.position(t2));
   std::vector<relation_t> extra({std::make_pair(w1, w2)});
-  Congruence              cong("twosided", &S, extra);
+  Congruence<>            cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 21);
@@ -231,7 +232,7 @@ TEST_CASE("Congruence 8L: left congruence on transformation semigroup size 88",
   // REQUIRE(S.degree() == 5);
   std::vector<relation_t> extra(
       {relation_t({0, 1, 0, 0, 0, 1, 1, 0, 0}, {1, 0, 0, 0, 1})});
-  Congruence cong("left", &S, extra);
+  Congruence<> cong("left", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 69);
@@ -282,7 +283,7 @@ TEST_CASE("Congruence 8R: right congruence on transformation semigroup size 88",
   S.factorisation(w1, S.position(t1));
   S.factorisation(w2, S.position(t2));
   std::vector<relation_t> extra({std::make_pair(w1, w2)});
-  Congruence              cong("right", &S, extra);
+  Congruence<>            cong("right", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 72);
@@ -331,7 +332,7 @@ TEST_CASE("Congruence 09: for an infinite fp semigroup",
                                        relation_t({1, 2}, {1}),
                                        relation_t({2, 2}, {1})};
   std::vector<relation_t> extra = {relation_t({0}, {1})};
-  Congruence              cong("twosided", 3, relations, extra);
+  Congruence<>            cong("twosided", 3, relations, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.word_to_class_index({0}) == cong.word_to_class_index({1}));
@@ -355,7 +356,7 @@ TEST_CASE("Congruence 10: for an infinite fp semigroup",
                                   relation_t({1, 2}, {1}),
                                   relation_t({2, 1}, {1})};
   std::vector<relation_t> extra = {{{0}, {1}}};
-  Congruence              cong("twosided", 3, rels, extra);
+  Congruence<>            cong("twosided", 3, rels, extra);
   cong.set_report(CONG_REPORT);
   // This line is here to make sure that the max_threads is ignored here, since
   // if we are limited to one thread here then this example doesn't run!
@@ -390,7 +391,7 @@ TEST_CASE("Congruence 11: congruence on big finite semigroup",
 
   std::vector<relation_t> extra(
       {relation_t({0, 3, 2, 1, 3, 2, 2}, {3, 2, 2, 1, 3, 3})});
-  Congruence cong("twosided", &S, extra);
+  Congruence<> cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.word_to_class_index({0, 0, 0, 1})
@@ -449,7 +450,7 @@ TEST_CASE("Congruence 12: Congruence on full PBR monoid on 2 points",
   std::vector<relation_t> extra(
       {relation_t({7, 10, 9, 3, 6, 9, 4, 7, 9, 10}, {9, 3, 6, 6, 10, 9, 4, 7}),
        relation_t({8, 7, 5, 8, 9, 8}, {6, 3, 8, 6, 1, 2, 4})});
-  Congruence cong("twosided", &S, extra);
+  Congruence<> cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 19009);
@@ -501,7 +502,7 @@ TEST_CASE("Congruence 13: partial perm example",
 
   std::vector<relation_t> extra
       = {relation_t({5, 3, 1}, {3, 3}), relation_t({2, 7}, {1, 6, 6, 1})};
-  Congruence cong("twosided", &S, extra);
+  Congruence<> cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 32);
@@ -516,7 +517,7 @@ TEST_CASE("Congruence 14: Bicyclic monoid",
                                   relation_t({2, 0}, {2}),
                                   relation_t({1, 2}, {0})};
   std::vector<relation_t> extra;
-  Congruence              cong("twosided", 3, rels, extra);
+  Congruence<>            cong("twosided", 3, rels, extra);
   cong.set_report(CONG_REPORT);
   REQUIRE(cong.word_to_class_index({0})
           == cong.word_to_class_index({1, 2, 1, 1, 2, 2}));
@@ -536,7 +537,7 @@ TEST_CASE("Congruence 15: Congruence on bicyclic monoid",
                                   relation_t({2, 0}, {2}),
                                   relation_t({1, 2}, {0})};
   std::vector<relation_t> extra({relation_t({1, 1, 1}, {0})});
-  Congruence              cong("twosided", 3, rels, extra);
+  Congruence<>            cong("twosided", 3, rels, extra);
 
   cong.set_report(CONG_REPORT);
 
@@ -553,7 +554,7 @@ TEST_CASE("Congruence 16: Congruence on free abelian monoid with 15 classes",
                                   relation_t({1, 2}, {2, 1})};
   std::vector<relation_t> extra(
       {relation_t({1, 1, 1, 1, 1}, {1}), relation_t({2, 2, 2}, {2})});
-  Congruence cong("twosided", 3, rels, extra);
+  Congruence<> cong("twosided", 3, rels, extra);
 
   cong.set_report(CONG_REPORT);
 
@@ -587,7 +588,7 @@ TEST_CASE("Congruence 17: Congruence on full PBR monoid on 2 points (max 2)",
   std::vector<relation_t> extra(
       {relation_t({7, 10, 9, 3, 6, 9, 4, 7, 9, 10}, {9, 3, 6, 6, 10, 9, 4, 7}),
        relation_t({8, 7, 5, 8, 9, 8}, {6, 3, 8, 6, 1, 2, 4})});
-  Congruence cong("twosided", &S, extra);
+  Congruence<> cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
   cong.set_max_threads(2);
 
@@ -645,7 +646,7 @@ TEST_CASE("Congruence 18: Congruence on full PBR monoid on 2 points (max 1)",
   std::vector<relation_t> extra(
       {relation_t({7, 10, 9, 3, 6, 9, 4, 7, 9, 10}, {9, 3, 6, 6, 10, 9, 4, 7}),
        relation_t({8, 7, 5, 8, 9, 8}, {6, 3, 8, 6, 1, 2, 4})});
-  Congruence cong("twosided", &S, extra);
+  Congruence<> cong("twosided", &S, extra);
   cong.set_report(CONG_REPORT);
   cong.set_max_threads(1);
 
@@ -692,7 +693,7 @@ TEST_CASE("Congruence 19: Infinite fp semigroup from GAP library",
                                   relation_t({2, 1}, {1})};
   std::vector<relation_t> extra = {relation_t({0}, {1})};
 
-  Congruence cong("twosided", 3, rels, extra);
+  Congruence<> cong("twosided", 3, rels, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(!cong.is_done());
@@ -710,7 +711,7 @@ TEST_CASE("Congruence 20: Infinite fp semigroup with infinite classes",
   std::vector<relation_t> rels
       = {relation_t({0, 0, 0}, {0}), relation_t({0, 1}, {1, 0})};
   std::vector<relation_t> extra = {relation_t({0}, {0, 0})};
-  Congruence              cong("twosided", 2, rels, extra);
+  Congruence<>            cong("twosided", 2, rels, extra);
   cong.set_report(CONG_REPORT);
 
   word_t x = {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1};
@@ -736,7 +737,7 @@ TEST_CASE("Congruence 21: trivial cong on an fp semigroup",
   rels.push_back(relation_t({0}, {1}));        // (a, b)
   std::vector<relation_t> extra;
 
-  Congruence cong("left", 2, rels, extra);
+  Congruence<> cong("left", 2, rels, extra);
 
   Partition<word_t>* ntc = cong.nontrivial_classes();
   REQUIRE(ntc->size() == 0);
@@ -753,7 +754,7 @@ TEST_CASE("Congruence 22: duplicate generators on a finite semigroup",
   Semigroup<> S = Semigroup<>(gens);
   S.set_report(CONG_REPORT);
   really_delete_cont(gens);
-  Congruence cong("twosided", &S, std::vector<relation_t>());
+  Congruence<> cong("twosided", &S, std::vector<relation_t>());
 
   REQUIRE(cong.nr_classes() == S.size());
 }
@@ -776,7 +777,7 @@ TEST_CASE("Congruence 23: test nontrivial_classes for a fp semigroup cong",
          relation_t({1, 1, 1, 1, 0, 1, 0}, {1, 0, 1, 0}),
          relation_t({0, 0, 1, 1, 1, 0, 1, 0}, {1, 1, 1, 0, 1, 0})};
 
-  Congruence cong(
+  Congruence<> cong(
       "twosided", 2, rels, std::vector<relation_t>({relation_t({0}, {1})}));
   cong.set_report(CONG_REPORT);
 
@@ -825,7 +826,7 @@ TEST_CASE("Congruence 24: example from GAP which once messed up prefill",
     delete elms[i + 1];
   }
 
-  Congruence cong("right", &S, extra);
+  Congruence<> cong("right", &S, extra);
   cong.set_report(CONG_REPORT);
 
   REQUIRE(cong.nr_classes() == 1);
@@ -834,31 +835,31 @@ TEST_CASE("Congruence 24: example from GAP which once messed up prefill",
 TEST_CASE("Congruence 25: free semigroup with redundant relations",
           "[quick][congruence][multithread][fpsemigroup][25]") {
   std::vector<relation_t> extra = {relation_t({0, 0}, {0, 0})};
-  Congruence              cong("twosided", 1, {}, extra);
+  Congruence<>            cong("twosided", 1, {}, extra);
   REQUIRE(cong.test_equals({0, 0}, {0, 0}));
 }
 
 TEST_CASE("Congruence 26: free semigroup with redundant relations",
           "[quick][congruence][multithread][fpsemigroup][26]") {
-  Congruence cong("twosided", 1, {}, {});
+  Congruence<> cong("twosided", 1, {}, {});
   REQUIRE(cong.test_equals({0, 0}, {0, 0}));
   REQUIRE(!cong.test_equals({0, 0}, {0}));
 }
 
 TEST_CASE("Congruence 27: is_obviously_infinite",
           "[quick][congruence][fpsemigroup][27]") {
-  Congruence cong1(
+  Congruence<> cong1(
       "twosided", 3, {relation_t({0, 1}, {0})}, {relation_t({2, 2}, {2})});
   REQUIRE(cong1.is_obviously_infinite());
-  Congruence cong2("twosided",
-                   3,
-                   {relation_t({0, 1}, {0}), relation_t({0, 0}, {0})},
-                   {relation_t({1, 1}, {1})});
+  Congruence<> cong2("twosided",
+                     3,
+                     {relation_t({0, 1}, {0}), relation_t({0, 0}, {0})},
+                     {relation_t({1, 1}, {1})});
   REQUIRE(cong2.is_obviously_infinite());
-  Congruence cong3("twosided",
-                   3,
-                   {relation_t({0, 1}, {0}), relation_t({0, 0}, {0})},
-                   {relation_t({1, 2}, {1})});
+  Congruence<> cong3("twosided",
+                     3,
+                     {relation_t({0, 1}, {0}), relation_t({0, 0}, {0})},
+                     {relation_t({1, 2}, {1})});
   REQUIRE(!cong3.is_obviously_infinite());
 
   std::vector<Element*> gens = {new Transformation<u_int16_t>({0, 1, 0}),
@@ -868,12 +869,43 @@ TEST_CASE("Congruence 27: is_obviously_infinite",
   S.set_report(CONG_REPORT);
   really_delete_cont(gens);
 
-  Congruence cong4("twosided", &S, {relation_t({1}, {0})});
+  Congruence<> cong4("twosided", &S, {relation_t({1}, {0})});
   REQUIRE(!cong4.is_obviously_infinite());
 }
 
 TEST_CASE("Congruence 28: test_less_than",
           "[quick][congruence][fpsemigroup][28]") {
-  Congruence cong("twosided", 2, {relation_t({0, 0}, {0})}, {});
+  Congruence<> cong("twosided", 2, {relation_t({0, 0}, {0})}, {});
   REQUIRE(!cong.test_less_than({0, 0}, {0}));
+}
+
+TEST_CASE("Congruence 29: test BMat8", "[quick][congruence][fpsemigroup][29]") {
+  std::vector<BMat8> gens
+      = {BMat8({{0, 1, 0, 0}, {1, 0, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 1}}),
+         BMat8({{0, 1, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 1}, {1, 0, 0, 0}}),
+         BMat8({{1, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 1, 0}, {1, 0, 0, 1}}),
+         BMat8({{1, 0, 0, 0}, {0, 1, 0, 0}, {0, 0, 1, 0}, {0, 0, 0, 0}})};
+
+  Semigroup<BMat8> S(gens);
+  S.set_report(CONG_REPORT);
+
+  Congruence<BMat8> cong1("twosided", &S, {relation_t({1}, {0})});
+  REQUIRE(cong1.nr_classes() == 3);
+  REQUIRE(cong1.word_to_class_index({1}) == 0);
+
+  Congruence<BMat8> cong2("twosided", &S, {relation_t({1}, {0})});
+  cong2.force_kbfp();
+  REQUIRE(cong2.nr_classes() == 3);
+  REQUIRE(cong2.word_to_class_index({1}) == 0);
+
+  Semigroup<BMat8>  T({gens[0], gens[2], gens[3]});
+  Congruence<BMat8> cong3("twosided", &T, {relation_t({1}, {0})});
+  cong3.force_p();
+  REQUIRE(cong3.nr_classes() == 2);
+  REQUIRE(cong3.word_to_class_index({1}) == 0);
+
+  Congruence<BMat8> cong4("twosided", &S, {relation_t({1}, {0})});
+  cong2.force_tc();
+  REQUIRE(cong4.nr_classes() == 3);
+  REQUIRE(cong4.word_to_class_index({1}) == 0);
 }
