+++ tests/rwse.test.cc
@@ -38,9 +38,12 @@ TEST_CASE("RWSE 01:", "[quick][rwse][01]") {
   REQUIRE(S.degree() == 2);
   REQUIRE(S.nrrules() == 4);
   std::vector<relation_t> extra;
-  Congruence              cong("twosided", &S, extra);
+  Congruence<>            cong("twosided", &S, extra);
 
-  RWS rws(cong);
+  RWS rws;
+  rws.add_rules(cong.relations());
+  rws.add_rules(cong.extra());
+
   REQUIRE(rws.confluent());
 
   gens          = {new RWSE(rws, 0), new RWSE(rws, 1)};
@@ -72,8 +75,11 @@ TEST_CASE("RWSE 02: factorisation", "[quick][rwse][02]") {
   really_delete_cont(gens);
 
   std::vector<relation_t> extra;
-  Congruence              cong("twosided", &S, extra);
-  RWS                     rws(cong);
+  Congruence<>            cong("twosided", &S, extra);
+  RWS                     rws;
+  rws.add_rules(cong.relations());
+  rws.add_rules(cong.extra());
+
   REQUIRE(rws.confluent());
 
   gens          = {new RWSE(rws, 0), new RWSE(rws, 1)};
